requirements.txt
# Flask ve web
Flask==3.1.0
Jinja2==3.1.5
Werkzeug==3.1.3

# Ortam değişkenleri
python-dotenv==1.0.1

# Veri işlemleri
pandas==2.2.3
openpyxl==3.1.5

# Veritabanı
PyMySQL==1.1.1

# Şifreleme / güvenlik
cryptography==43.0.3
bcrypt==4.2.1

# E-posta doğrulama
email-validator==2.2.0
dnspython==2.7.0
python-whois==0.9.5  # Opsiyonel: domain yaşı kontrolü için

# HTTP / API
requests==2.32.3

# AWS SES için
boto3==1.35.99
